requirements.txt
PyPDF2==3.0.1
PyMuPDF==1.23.7
Pillow==10.1.0
pytesseract==0.3.10
pdf2image==1.16.3
opencv-python==4.8.1.78
numpy>=1.26.0